+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to `@homebridge/plugin-ui-utils` will be documented in this 
 
 - chore: dependency updates
 - docs: add `CLAUDE.md` for claude code
+- fix(ui): verify origin and source on incoming postMessage
 
 ## v2.2.3 (2026-03-29)